time-dependent Liouvillian support, and options in ODE solver
Made a few adjustments, only to the bosonic solver at this stage:

1)  Added support for time-dependent Liouvillians.  In other words,
instead of
[H0, [H1, drive()]]

One can send

[L_0, [L_1, drive()]]

where L_0 and L_0 are arbitrary superoperators (ideally liouvillians!)

2) added some boolean options in .run().  These allow for the initial
condition to be a numpy array containing the state of all ADOs, and
addiitionally for the ODE solver to return a array of ADOs (reshaped).
These can then be used with the ancilla functions in the notebooks to
get properties of the bath.